Major structural changes in the repo. Simulations:   Monte Carlo and MetropolisMontecarlo Done.   ToDO: make simulations nparticle insensitiv.   implement lambda, eds system
