+++ README.md
@@ -1,2 +1,5 @@
 # ConveyorBelt
-Python implementation of the conveyor belt scheme.
+Python implementation of the conveyor belt scheme and more.
+
+##Useage
+If you want to include this package into your scripts, source the package and import main or the package folder.
