+++ main.py
@@ -1,0 +1,3 @@
+from src import conveyorBelt, ensemble, integrator, potential1D, system
+import plots
+
