+++ src/ensemble.py
@@ -1,0 +1,5 @@
+"""
+Module: ensemble
+    This module shall be used to implement subclasses of ensemble.
+    It is a class, that is using multiple system. It can be used for RE or Conveyor belt
+"""
