+++ test/test_conveyorBelt.py
@@ -1,6 +1,5 @@
 import unittest
-import system1D
-import conveyorBelt
+from src import system, conveyorBelt
 
 
 class testEnsemble(unittest.TestCase):
@@ -12,39 +11,39 @@ class testEnsemble(unittest.TestCase):
         ens.print_systems()
 
     def testEnsembleSystem(self):
-        ens = conveyorBelt.Ensemble(0.0, 1, system=system1D.system1D(temp=300.0,
-                                                                     fc=1.0,
-                                                                     lam=0.5,
-                                                                     alpha=10.0,
-                                                                     integrator='sd'))
+        ens = conveyorBelt.Ensemble(0.0, 1, system=system.system1D(temp=300.0,
+                                                                   fc=1.0,
+                                                                   lam=0.5,
+                                                                   alpha=10.0,
+                                                                   integrator='sd'))
         ens.calc_ene()
         ens.propagate()
         ens.calc_ene()
         ens.print_systems()
 
     def testEnsembleSystemShift(self):
-        ens = conveyorBelt.Ensemble(0.0, 1, system=system1D.system1D(temp=300.0,
-                                                                     fc=1.0,
-                                                                     lam=0.5,
-                                                                     alpha=10.0,
-                                                                     integrator='sd'))
+        ens = conveyorBelt.Ensemble(0.0, 1, system=system.system1D(temp=300.0,
+                                                                   fc=1.0,
+                                                                   lam=0.5,
+                                                                   alpha=10.0,
+                                                                   integrator='sd'))
         ens.calc_ene()
         ens.propagate()
         ens.calc_ene()
         ens.print_systems()
 
     def testTraj(self):
-        ens = conveyorBelt.Ensemble(0.0, 1, system=system1D.system1D(temp=300.0,
-                                                                     fc=1.0,
-                                                                     lam=0.5,
-                                                                     alpha=10.0,
-                                                                     integrator='sd'))
+        ens = conveyorBelt.Ensemble(0.0, 1, system=system.system1D(temp=300.0,
+                                                                   fc=1.0,
+                                                                   lam=0.5,
+                                                                   alpha=10.0,
+                                                                   integrator='sd'))
         print(conveyorBelt.calc_traj(steps=10, ens=ens))
-        ens = conveyorBelt.Ensemble(0.0, 8, system=system1D.system1D(temp=300.0,
-                                                                     fc=1.0,
-                                                                     lam=0.5,
-                                                                     alpha=10.0,
-                                                                     integrator='sd'))
+        ens = conveyorBelt.Ensemble(0.0, 8, system=system.system1D(temp=300.0,
+                                                                   fc=1.0,
+                                                                   lam=0.5,
+                                                                   alpha=10.0,
+                                                                   integrator='sd'))
         conveyorBelt.calc_traj_file(steps=100, ens=ens)
