+++ test/test_system1D.py
@@ -1,6 +1,5 @@
 import unittest
-import potential1D as pot
-import system1D
+from src import system, potential1D as pot
 
 
 class testHarmonicOsc1D(unittest.TestCase):
@@ -12,8 +11,8 @@ class testHarmonicOsc1D(unittest.TestCase):
         fc = 1.0
         for fc in [1.0, 1.5, 10.0, 20.0]:
             with self.subTest(fc=fc):
-                sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                        potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
+                sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                                      potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
                 sys.pos = 1.0
                 sys.updateEne()
                 self.assertAlmostEqual(sys.totpot, fc)
@@ -29,8 +28,8 @@ class testHarmonicOsc1D(unittest.TestCase):
 
         for alpha in [1.0, 10.0, 100.0]:
             with self.subTest(alpha=alpha):
-                sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                        potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
+                sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                                      potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
                 sys.pos = 1.0
                 sys.updateEne()
                 self.assertAlmostEqual(sys.totpot, 0.5 + 0.5 * alpha)
@@ -46,8 +45,8 @@ class testHarmonicOsc1D(unittest.TestCase):
 
         for lam in [0.0, 0.5, 1.0]:
             with self.subTest(lam=lam):
-                sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                        potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
+                sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                                      potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
                 sys.pos = 1.0
                 sys.updateEne()
                 self.assertAlmostEqual(sys.totpot, 0.5 + 0.5 * lam)
@@ -66,8 +65,8 @@ class testHarmonicOsc1D(unittest.TestCase):
                 for sd in [0, 1]:
                     for nose in [0, 1]:
                         for mass in [[1.0, 1.0], [1.0, 2.0], [2.0, 2.0]]:
-                            sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                                    potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
+                            sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                                                  potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
 
     def testNoseHoover(self):
         temp = 300.0
@@ -77,8 +76,8 @@ class testHarmonicOsc1D(unittest.TestCase):
         fc = 1.0
         for vel in [1.0, 2.0, 3.0]:
             with self.subTest(vel=vel):
-                sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                        potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
+                sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                                      potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha))
                 sys.vel = vel
                 sys.updateEne()
                 self.assertAlmostEqual(sys.totkin, vel ** 2 / 4.0)
@@ -92,8 +91,8 @@ class testHarmonicOsc1D(unittest.TestCase):
         fc = 1.0
         for vel in [1.0, 2.0, 3.0]:
             with self.subTest(vel=vel):
-                sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                        potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha), integrator='newton')
+                sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                                      potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha), integrator='newton')
                 sys.pos = 1.0
                 sys.newpos = 1.0
                 sys.newvel = vel
@@ -105,8 +104,8 @@ class testHarmonicOsc1D(unittest.TestCase):
                 self.assertAlmostEqual(sys.vel, vel - 0.2)
                 self.assertAlmostEqual(sys.newpos, 1.0 + (vel - 0.4) * 0.1)
         vel = 1.0
-        sys = sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                      potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha), integrator='newton')
+        sys = sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                                    potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha), integrator='newton')
         sys.pos = 1.0
         sys.newpos = 1.0
         sys.vel = 1.0
@@ -122,8 +121,8 @@ class testHarmonicOsc1D(unittest.TestCase):
         alpha = 1.0
         lam = 1.0
         fc = 1.0
-        sys = system1D.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
-                                potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha), integrator='hmc')
+        sys = system.system1D(temp=temp, fc=fc, lam=lam, alpha=alpha, mass=mass,
+                              potential=pot.pertHarmonicOsc1D(fc=fc, alpha=alpha), integrator='hmc')
 
 
 if __name__ == '__main__':
